+++ src/cmigits/cmigitsDaemon.cpp
@@ -159,16 +159,16 @@ main(int argc, char *argv[]) {
 
     // When new status arrives, stuff into shared memory.
     QObject::connect(
-            Cm, SIGNAL(new3500Data(uint64_t, uint16_t, bool, bool, bool, uint16_t, uint16_t, uint16_t, uint16_t, uint16_t, float, float, float)),
-            &shm, SLOT(storeLatest3500Data(uint64_t, uint16_t, bool, bool, bool, uint16_t, uint16_t, uint16_t, uint16_t, uint16_t, float, float, float)));
+            Cm, SIGNAL(new3500Data(uint64_t, uint16_t, bool, bool, bool, uint16_t, uint16_t, uint16_t, uint16_t, uint16_t, double, double, double)),
+            &shm, SLOT(storeLatest3500Data(uint64_t, uint16_t, bool, bool, bool, uint16_t, uint16_t, uint16_t, uint16_t, uint16_t, double, double, double)));
     // When new navigation solution arrives, stuff into shared memory.
     QObject::connect(
-            Cm, SIGNAL(new3501Data(uint64_t, float, float, float, float, float, float)),
-            &shm, SLOT(storeLatest3501Data(uint64_t, float, float, float, float, float, float)));
+            Cm, SIGNAL(new3501Data(uint64_t, double, double, double, double, double, double)),
+            &shm, SLOT(storeLatest3501Data(uint64_t, double, double, double, double, double, double)));
     // When new attitude arrives, stuff into shared memory.
     QObject::connect(
-            Cm, SIGNAL(new3512Data(uint64_t, float, float, float)),
-            &shm, SLOT(storeLatest3512Data(uint64_t, float, float, float)));
+            Cm, SIGNAL(new3512Data(uint64_t, double, double, double)),
+            &shm, SLOT(storeLatest3512Data(uint64_t, double, double, double)));
 
     // Create our XML-RPC method registry and server instance
     PMU_auto_register("instantiating XML-RPC server");
